src/backend/requirements.txt
greenlet==2.0.2
mysqlclient==2.1.1
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.2
six==1.16.0
SQLAlchemy==2.0.7
typing_extensions==4.5.0
